Fix `non-serializable value in the state` error
There was an error in the console when clicking `Generate deposit address` on
tbtc page:

`A non-serializable value was detected in the state`

This happened because we passed non-serializable values (in this case functions
- `handleJsonDownload` and `handleDoubleReject`) as a parameters throug a redux
action.

This commit moves those two function straight into the `TbtcRecoveryJson` modal.
