+++ src/pages/tBTC/Bridge/MintingCard/ProvideData.tsx
@@ -75,23 +75,7 @@ const MintingProcessForm = withFormik<MintingProcessFormProps, FormValues>({
 export const ProvideData: FC = () => {
   const { updateState, ethAddress, btcRecoveryAddress } = useTbtcState()
   const formRef = useRef<FormikProps<FormValues>>(null)
-  const { openModal, closeModal } = useModal()
-
-  const handleJsonDownload = (data: DepositScriptParameters) => {
-    downloadFile(
-      JSON.stringify(data),
-      "deposit-script-parameters.json",
-      "text/json"
-    )
-
-    closeModal()
-    updateState("mintingStep", MintingStep.Deposit)
-  }
-
-  const handleDoubleReject = () => {
-    updateState("mintingStep", MintingStep.Deposit)
-    closeModal()
-  }
+  const { openModal } = useModal()
 
   const onSubmit = async (values: FormValues) => {
     // check if the user has changed the eth or btc address from the previous attempt
@@ -126,8 +110,6 @@ export const ProvideData: FC = () => {
       // if the user has NOT declined the json file, ask the user if they want to accept the new file
       openModal(ModalType.TbtcRecoveryJson, {
         jsonData: depositScriptParameters,
-        handleDownloadClick: handleJsonDownload,
-        handleDoubleReject,
       })
     }
